@@ -12,10 +12,10 @@
     "express-jwt": "^6.0.0",
     "jsonwebtoken": "^8.5.1",
     "jwt-decode": "^3.1.2",
-    "mdbreact": "^5.0.1",
-    "react": "^18.2.0",
+    "mdbreact": "^5.2.0",
+    "react": "^18.3.1",
     "react-countdown": "^2.3.1",
-    "react-dom": "^18.2.0s",
+    "react-dom": "^18.3.1",
     "react-router-dom": "^5.2.0",
     "react-scripts": "5.0.1",
     "socket.io-client": "^4.0.1",
